+++ src/integrity/mod.rs
@@ -2,6 +2,7 @@
 mod checks;
 pub mod error;
 mod issue;
+pub mod metadata;
 mod repairer;
 mod utils;
 mod verifier;